rename rubo-cop to rubocop
