+++ src/linters/rubocop.js
