+++ test/linters/projects/rubocop/file1.rb
