+++ test/linters/projects/rubocop/file2.rb
